pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "diffusion-agent"
version = "0.1.0"
description = "Agent-integrated system for enabling PyTorch model support on Huawei Ascend NPU"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
dependencies = [
    "langgraph>=0.2.0",
    "langchain-core>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-anthropic>=0.2.0",
    "typer>=0.12.0",
    "pydantic-settings>=2.0.0",
    "structlog>=24.0.0",
    "pyyaml>=6.0",
    "jinja2>=3.1.0",
    "gitpython>=3.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-mock>=3.12",
    "ruff>=0.5.0",
]

[project.scripts]
diffusion-agent = "diffusion_agent.cli:app"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.pytest.ini_options]
testpaths = ["tests"]
